+++ raiden-dapp/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - [#694] Show protocol progress steps for transfers.
+- [#687] Ghost action buttons.
 
 ### Fixed
 - [#1243] Fix width of transfer input on mobile screens.
@@ -12,6 +13,7 @@
 [#694] https://github.com/raiden-network/light-client/issues/694
 [#1243]: https://github.com/raiden-network/light-client/issues/1243
 [#1242]: https://github.com/raiden-network/light-client/issues/1242
+[#687] https://github.com/raiden-network/light-client/issues/687
 
 ## [0.5.0] - 2020-03-27
 ### Added